Update zh_cn.lang
I posted a version of translation about one month ago and sent it via curseforge to UnlimitedStone9, but didn't pull request on Github. HeartyYF has invited me to the TFC Chinese translation group before and recently checked the current translation file on GitHub, found it's mostly machine translation, so he contacted me to fix this. Here I upload my version of the translation which is consistent with the main TFCmod translation, and will keep updating it